pyproject.toml
[project]
name = "ai-agent-traders"
version = "0.1.0"
description = "Autonomous AI Trading Team powered by OpenClaw"
authors = [
    { name = "Antigravity", email = "antigravity@example.com" }
]
dependencies = [
    "ccxt>=4.2.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.6.0",
    "ruff>=0.3.0",
]
requires-python = ">=3.10"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL"]
ignore = ["PLR2004", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
